js для изменения кол-ва товаров в корзине без перезагрузки страницы

+++ store/views.py
@@ -2,34 +2,25 @@ import json
 
 from django.contrib.auth.decorators import login_required
 from django.http import JsonResponse
-from django.shortcuts import render, redirect
-from django.views.generic import ListView, DetailView
+from django.shortcuts import render
+from django.views.generic import ListView
 
-from .models import Product, Order, OrderItem
+from .models import Product, Order, OrderItem, DeliveryDate
 
 
 class ProductListView(ListView):
     model = Product
     template_name = 'store/index.html'
 
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['order'] = False
-        if self.request.user.is_authenticated:
-            if Order.objects.filter(customer=self.request.user,
-                                    status='formed').exists():
-                context['order'] = True
-        return context
-
 
 def basket(request):
     if request.user.is_authenticated:
-        if Order.objects.filter(customer=request.user,
-                                status='formed').exists():
-            return redirect('orders')
         order, created = Order.objects.get_or_create(
             customer=request.user,
             status='not formed')
+        if DeliveryDate.objects.filter(status='open').exists():
+            order.delivery_date = DeliveryDate.objects.get(status='open')
+            order.save()
         items = order.orderitem_set.all()
     else:
         items = []
@@ -52,25 +43,25 @@ def about(request):
 
 
 def update_item(request):
-    data = json.loads(request.body)
-    product_id = data['productId']
-    action = data['action']
-
-    product = Product.objects.get(id=product_id)
-    order, created = Order.objects.get_or_create(customer=request.user,
-                                                 status='not formed')
-    order_item, created = OrderItem.objects.get_or_create(order=order,
-                                                          product=product)
-    if action == 'add':
-        order_item.quantity += 1
-    elif action == 'remove':
-        order_item.quantity -= 1
-
-    order_item.save()
-    if order_item.quantity <= 0:
-        order_item.delete()
-
-    return JsonResponse('Item was added', safe=False)
+    if request.method == 'POST':
+        data = json.loads(request.body)
+        product_id = int(data['productId'])
+        action = data['action']
+        product = Product.objects.get(id=product_id)
+        order, created = Order.objects.get_or_create(customer=request.user,
+                                                     status='not formed')
+        order_item, created = OrderItem.objects.get_or_create(order=order,
+                                                              product=product)
+        if action == 'add':
+            order_item.quantity += 1
+        elif action == 'remove':
+            order_item.quantity -= 1
+        order_item.save()
+        if order_item.quantity <= 0:
+            order_item.delete()
+        return JsonResponse({"quantity": order_item.quantity,
+                             "total_items": order.get_total_items}, status=200)
+    return JsonResponse('GET method not allowed', safe=False)
 
 
 def make_order(request):
